relocate METAMASK_CONTROLLER_EVENTS
