+++ app/scripts/controllers/app-state.js
@@ -1,6 +1,6 @@
 import EventEmitter from 'events';
 import { ObservableStore } from '@metamask/obs-store';
-import { METAMASK_CONTROLLER_EVENTS } from '../metamask-controller';
+import { METAMASK_CONTROLLER_EVENTS } from '../constants/events.js';
 import { MINUTE } from '../../../shared/constants/time';
 
 export default class AppStateController extends EventEmitter {
