+++ app/scripts/lib/personal-message-manager.js
@@ -4,7 +4,7 @@ import { bufferToHex, stripHexPrefix } from 'ethereumjs-util';
 import { ethErrors } from 'eth-rpc-errors';
 import log from 'loglevel';
 import { MESSAGE_TYPE } from '../../../shared/constants/app';
-import { METAMASK_CONTROLLER_EVENTS } from '../metamask-controller';
+import { METAMASK_CONTROLLER_EVENTS } from '../constants/events.js';
 import createId from '../../../shared/modules/random-id';
 import { EVENT } from '../../../shared/constants/metametrics';
 import { addHexPrefix } from './util';
