+++ app/scripts/metamask-controller.js
@@ -144,14 +144,6 @@ import {
 } from './controllers/permissions';
 import createRPCMethodTrackingMiddleware from './lib/createRPCMethodTrackingMiddleware';
 
-export const METAMASK_CONTROLLER_EVENTS = {
-  // Fired after state changes that impact the extension badge (unapproved msg count)
-  // The process of updating the badge happens in app/scripts/background.js.
-  UPDATE_BADGE: 'updateBadge',
-  // TODO: Add this and similar enums to @metamask/controllers and export them
-  APPROVAL_STATE_CHANGE: 'ApprovalController:stateChange',
-};
-
 export default class MetamaskController extends EventEmitter {
   /**
    * @param {Object} opts
